Don't pin launcher app in memory
Similar to what we did for the camera app, unpin the launcher app from
memory as well. While the default launcher (Launcher3) isn't
particularly big, it doesn't make much sense to pin because the launcher
does not typically load new resources much. Most of its resources should
already be loaded in memory after it starts, so pinning the APK is
redundant.

Change-Id: Ia03de9c546a042637647f45d7b2b4072c1555fb5
Reviewed-on: https://review.statixos.com/c/android_device_google_gs101/+/8279
Reviewed-by: Anay Wadhera <anay1018@gmail.com>
Tested-by: Anay Wadhera <anay1018@gmail.com>

# Conflicts:
#	overlay/frameworks/base/core/res/res/values/config.xml

Change-Id: I8e2d04a0bc2d9dce4c5c2e8904753df2c37ec103
Signed-off-by: Matt Filetto <matt.filetto@gmail.com>